refactor: reorganize modules and extract timer utilities
- addTask(name, estimatedPomos) — silently dropped project/priority/category
- Move notification store to features/notifications for better organization
- Extract timer input utilities into dedicated timer-utils module
- Move category breakdown UI to reusable component
- Consolidate duration state management in settings focus section
- Update import paths to reflect new module structure
